+++ packages/slonik-interceptor-field-name-transformation/CHANGELOG.md
@@ -1,5 +1,7 @@
 # slonik-interceptor-field-name-transformation
 
+## 48.13.0
+
 ## 48.12.3
 
 ## 48.12.2