+++ packages/slonik-sql-tag-raw/CHANGELOG.md
@@ -1,5 +1,12 @@
 # slonik-sql-tag-raw
 
+## 48.13.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @slonik/sql-tag@48.13.0
+
 ## 48.12.3
 
 ### Patch Changes