+++ packages/sql-tag/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @slonik/sql-tag
 
+## 48.13.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @slonik/errors@48.13.0
+  - @slonik/types@48.13.0
+
 ## 48.12.3
 
 ### Patch Changes